Remove kotlin-android-extensions plugin
The plugin has been deprecated since Kotlin 1.4.20
(https://android-developers.googleblog.com/2020/11/the-future-of-kotlin-android-extensions.html)
and is set to be removed in Kotlin 1.8 (which according to the blog post
https://android-developers.googleblog.com/2022/02/discontinuing-kotlin-synthetics-for-views.html
is set to be released in end-2022).

View Binding is recommended as a substitute for Kotlin Synthetics, and
as the last change to use View Binding was done in the previous commit
(2b957b938a717dd59757d644b949459c6d8f2948), it should hence be safe to
remove the plugin with hopefully no issues.