ModalBottomSheetDSL: Fix group sometimes not applying in certain cases
This happens if the overload that uses a DSL for _both_ a group and its
items were to be used, as it does not also set the `group` on the
corresponding items.

A solution would to instead only set the group of each item when the
group items DSL's `build` method is called.

Also updates some KDocs to reflect the changes