TodoFragment: Specify the lifecycle owner for menu provider
The Javadoc for `addMenuProvider` states that:

> Adds the given MenuProvider to this MenuHost. If using this method,
> you must manually remove the provider when necessary.

As we're removing it manually, the app would crash if the user were to
navigate away from the fragment ("TodoFragment is not attached to a
fragment manager").

This change adds the missing lifecycle owner parameter which would allow
for the `MenuHost` to automatically remove the menu provider when the
owner is destroyed.