EditTaskActivity: Fix runtime crash
`launch` does **not** block any further execution, so the code right
below the code used to retrieve the task's metadata would cause a
runtime exception as the `lateinit` `todoItem` property was not set.

Fix this issue by moving the entire views-related logic to directly
after the task metadata retrieval.

Also fixes a UI issue where the progress bar UI was shown instead of the
task metadata, and updates the default visibility of the progress
bar/`ScrollView` accordingly in the layout XML.