Add extension function for `?.let { { MyComposable() } }`
Adds a `.letComposable {}` scope function which is similar to the `let`
scope function (https://kotlinlang.org/docs/scope-functions.html#let),
but returns a `@Composable () -> R` function.

This improves the readability for code that allows for a nullable
`@Composable () -> Unit` lambda to be used (which previously used
`value?.let { { MyComposable(value = it) } }`, which looks unreadable
with the extra curly brackets).