Introduce platform-agnostic User class
This removes the tight coupling to Firebase Auth
in case we do plan to migrate to another platform.

Also makes it easier to inject a "user" when writing
tests that require a currently signed-in user.

This commit only adds the initial `User` class and
an `AuthState` sealed interface, as well as mappers
to the class